Fix mk-string-arrays.awk to accept include file as parameter
Previews "Remove the uneeded err_detail_type.h ..." commit changed
mk-string-arrays.awk to add an include file with array definition
only when this array is an array of SBufs. The reasons are:
  a) I (wrongly?) believed that only in the case of SBus is needed.
  b) It computed the include file name using the enum name. In the case of
     error/categories.cc the include file (error/forward.h) does not match
     the enum name (err_type)

This patch fixes mk-string-arrays.awk to accept an include file name
as parameter and add it to #include files list if given. Use