Implement Ssl::PeerCertificatesVerify which verify Server certificates
The OpenSSL SSL_connect must be complete succesfully, and without found
any issues to certificates to be able to use this function with a SSL
object.
The major reason is that if the certificates verification failed
(while SSL_connect executed) the peer certificates are not attached
to SSL object and can not be retrieved with SSL_get_peer_cert_chain
function.