Handling missing issuer certificates for TLSv1.3, First implementation
This patch try to handle the missing issuer certificate problem
when the TLSv1.3 is used. When TLSv1.3 is used the server certificates
are sent encrypted to the client so existing squid code which parses
server hello, extract server certificates, check for missings
and download missing if required, can not work.

This patch when a missing certificate error is detected and if the
URI of missing certificate exist, ignore the error and schedules a
second certificates verify check for after SSL_connect is finished
and after the missing certificates are downloaded.

This patch try to use old code for handling missing certificates
it is operational but needs a lot of work to:
 - adjust old code for new needs
 - remove uneeded old code