Call neighbors_init() only after indexing completion (-F)
This high-level method eventually/asynchronously initiates
peers polling, which may start before the end of indexing.
For example, multicast peers (if configured) will be polled in a
10-second hard-coded delay, scheduled from peerDNSConfigure().
With '-F' and indexing in progress, we need to avoid this activity
because peers may incorrectly think that the Squid instance is ready
(which is not, in fact). We need to postpone neighbors_init()
in the same way as listeningPortsOpen() is.