Fixed commit 1ad6851 to delay buffer allocations until from-server reads
The commit made buffer allocations in maybeMakeSpaceAvailable() optional
so that they can be delayed until we actually read from the peer socket.
Delaying buffer allocation until read(2) is the right thing to do. The
callers (correctly) passed a boolean flag to trigger the allocation, but
the method interpreted the flag as if it meant "do not allocate". That
is, commit tried to delay allocations but actually allocated early.

Why was this bug unnoticed since 2015? The outside observers could not
see the problem because Squid continued to allocate the buffer, and
there was no code that would fiddle with that buffer while Squid was
waiting for the socket to become ready for reading (unlike bug 4206 that
served as an inspiration for commit 1ad6851).