Reworked readReply/maybeReadVirginBody() split, addressing many XXXs
Also do not wait for the rest of the HTTP response header or chunk
metadata if we have no space to store more incoming bytes. The parser
does not know how much more buffer space we have. It is our
responsibility to check that progress is possible.

I replaced "Non-HTTP-compliant" with "cannot parse" because we do not
know whether the parsing failure is related to HTTP compliance. For
example, buffer space limitations are not a compliance-related problem.

The changes near wasThereAnException look wrong, but it is the existing
wasThereAnException name that is "reversed". Compiled code should be OK.