United the 2-byte condition with the delayRead() workaround
... to make it easier to document it. This is a minor "fixup" change,
but the following related observation (and test results) are important.

The official code called delayRead() in HttpStateData::readReply(). The
latter method was probably not called when a full buffer was detected by
maybeReadVirginBody(), probably stalling the Squid-server transaction.
Moreover, if the buffer "became" full (due to configuration changes)
later, while we were waiting for the socket to become readable, then the
transaction would be killed by Must(maybeMakeSpaceAvailable(true))!

We are still not out of the woods yet: For example, setting
read_ahead_gap to 2 bytes stalls the transaction because the readGoal<=2
workaround defers the read before we have any body data to send to the
client. Without that data, nothing kicks our delayed read.