fixup: clarified why it is safe to downcast an uint64_t value to size_t
In general, unsigned downcast is _not_ safe. It is not a UB, but it will
blindly cut the highest bits off instead of returning smaller::max(). In
this particular case, we know that the value already fits into the
smaller/size_t type because it is a _minimum_ of such a value and X.