Improved StoreClient API const-correctness
A non-const isHit() testing method was a red flag, but the underlying
culprit was the StoreClient::loggingTags() method. That API should have
been declared const because it returned an essentially (and, in some
cases, literally) mutable history/statistics object. The fact that it
was possible to add "const" without any other modifications supports the
assertion that these methods naturally preserve their object state.