Reworked to a 'single timeout handler for all IpcIoFile objects' model
With the previous approach (when each IpcIoFile scheduled its own
open timeout) it was difficult to re-schedule these timeouts because
we cannot separate these events from each other in enentDelete()
(i.e., we cannot delete and add an event for a specific IpcIoFile).

Now these objects are stored in a map sorted by their expiration time
and there is a single scheduled event (at each moment) for the head
of the queue. The similar approach is used already in
the 'ping timeout service' implementation.