Fix Segfault while the PeerConnector SSL (OpenSSL) object is freed
The segfault happens while one of the X509 object which is attached to
the SSL object, is released. The X509 object appeared as an already
released object.

The issue exists inside squidX509VerifyCert function. Here a temporary
STACK_OF(X509) object is built (the untrustedCerts), which does not
lock the linked X509 objects but while it is destroyed (through the
Ssl::X509_STACK_Pointer destructor) at the end of the same function
will call X509_free for linked X509 objects.