fixup: Renamed JobCallbackPointer to JobWait, polished method names
... and synced the corresponding data member names.

The new class is not really a smart pointer (to a callback). With some
effort, it can be viewed as a smart pointer to a helper job, but that is
not really its primary purpose/role. The class is actually focusing on
the _wait_ (for our helper job calling our callback).

Renaming data members probably increases the diff, but the vast majority
of changes appear to be necessary for other reasons anyway.

Also swapped the JobWait::start() arguments for an arguably more natural
or "chronological" order.

Also found (two instances of) a branch bug in icap/Xaction.cc (now XXXs)
and sketched a TODO solution.