PeerConnector class: Replace gotServerCertificates() with a hasServerCertficates() method
We may have received the server certificates but maybe we do not have
access on them.

The old PeerConnector::gotServerCertificates() uses a flag which is
set in ssl_verify_cb function. This flag always set if we received
some server certificates from remote server.

The new PeerConnector::hasServerCertficates checks using the OpenSSL
function SSL_get_peer_cert_chain. This is may return nil if
OpenSSL verify procedure found one or more issues with certificates.

Because we are using SSL_get_peer_cert_chain to retrieve remote server
certificates the PeerConnector::hasServerCertficates is the right check
to see if we can procced with doValidationCallouts.