Disker mtRebuildFinished messages should affect newly registered workers
The problem with the mtRebuildFinished mechanism, added in 545c335, was
that in the case of small directories, diskers usually complete their
indexing before some workers register in Coordinator. Now Coordinator
maintains the list of already indexed diskers and notifies newly
registered workers with them.

Also fixed StoreController::FullyIndexed() by adding a missed
'Validated' condition.

Also fixed the timing of worker's "Finished rebuilding storage from
disk" event, which now occurs when it should - after all diskers
complete the indexing.

Also clarified the serverConnectionsOpen() logic in main.cc. This method
should be conditioned on useFullyIndexedStore() if and only if we are
forced to wait for indexing (-F applied and there are active dirs),
otherwise we should call it there and when it was called
previously (before the PR changes).