Simplify suspendNegotiation/resumeNegotiation mechanism
The PeerConnector::resumeNegotiationError renamed to resumeTlsNegotiationCb
and adjusted to be used as callback to resume Negotiation for all
cases we are suspending.

The NegotiationErrorDetails renamed to TlsNegotiationDetails. Now it
does not hold only error details but the result of the latest SSL_connect
call.

Also the folowing incosistency is fixed: After the missing certificates
are downloaded and peer certificates verified inside
missingCertificatesRetrieved method, in the case an error found
the old code call the noteNegotiationError leaving the PeerConnector
in suspending state.
This patch fixes this incosistency. Now always we are resuming even
after an error.