fixup: Simplify and protect private PeerConnectionPointer::connection_
... from being cleared by the (difficult-to-track) 3rd-party code.

AFAICT, these 3rd-party resets were (ab)used to prevent the pending
connection from being closed after a successful hand-off to a a helper
job. This commit essentially replaces

    try {
        ... step creation/scheduling code here ...
        conn = nullptr;
    } catch (...) {
        if (conn) // we still own the connection
            closePendingConnection(conn);
        else
            cancelStep();
    }

with the equivalent but much simpler

    try {
        ... step creation/scheduling code here ...
    } catch (...) {
        closePendingConnection(conn); // we still own the connection
    }

I think we can and should assert that the helper job is

* either scheduled to start (no exception is thrown in this case and the
  connection ownership passes to the helper job)

* or is not scheduled to start (an exception is thrown and the
  connection ownership stays with the caller).

The PR code assumed the existence of a third outcome: The helper job was
scheduled to start but the same code that successfully scheduled the job
to start threw an exception. I believe that if this outcome is possible,
then there is a bug somewhere outside of advanceDestination(). We should
fix any such bugs, but advanceDestination() should assume that there are
none because working around such bugs is too costly.

With this approach, we do not need to clear the conn pointer in the step
creation/scheduling code to signal a successful hand-off -- if we see
the exception, we know that we still own the connection and do not need
to cancel the step (which failed to start).

FWIW, PeerConnectionPointer::print(), at least, thinks that "we should
see no combinations of a nil connection and a set position". External
resets would invalidate that assumption. This is a secondary/minor
reason, of course.

Said that, I am pretty sure that the (future) connection ownership
object will be modified on the path modified by this commit because the
connection ownership does change hands on that path. Thus, we may have
to adjust advanceDestination() and step profiles again. Hopefully, by
the time we are ready for that adjustment, the code will not have to
sneak into PeerConnectionPointer private areas to pass conn ownership.