fixup: Remove unneeded code related to parsing certificates
I do not know why this code was not removed in branch commit bd4f158,
but I see no reason to fiddle with the details of certificate parsing if
Squid does not need to parse certificates.

AFAICT, the only downside of removing this code is that it changes the
absolute values of ParserState constants, but any related inconvenience
should be temporary, while the benefit of less code/ink is long-term. If
this is a more serious issue than I think, then we can preserve the old
values while removing atCertificatesReceived.