Always handle ignored X509_V_ERR_UNABLE_TO_GET_ISSUER_CERT_LOCALLY
Ignoring X509_V_ERR_UNABLE_TO_GET_ISSUER_CERT_LOCALLY in the middle of
SSL_connect() carries significant risks. We must always react to that
event, regardless of the SSL_connect() result. At the very least, we
must check that the negotiation has not completed successfully after we
hid X509_V_ERR_UNABLE_TO_GET_ISSUER_CERT_LOCALLY! I suspect the code
missed other (unusual but not theoretically impossible) cases as well.

The code now documents and, hopefully, handles all outcomes.

It may be tempting _not_ to fetch the missing certificates if the
negotiation has also failed for some other grave reason (and the
connections are not going to be spliced). However, we should resist this
temptation IMO because deciding whether the reason is grave or whether
the connections are going to be spliced may depend on what validation
errors Squid has encountered (via admin's ACLs). Hiding one of the
biggest errors when making those decisions feels too risky. Reporting
accurate trust validation results may also be important in error triage.

This commit documents and follows an "approximate the effects of a
built-in download of missing certificates" principle. The code tries to
mimic what would happen if OpenSSL could download the missing
certificates before (or while) validating the chain. Changes in download
failure reaction code may be needed to fully implement that principle.