Disker should always send mtRebuildFinished if it is already indexed
This should cover a case when a disker crashes just after indexing,
not having a chance to notify Coordinator. With this approach, the
notification will be sent twice if the previous disker instance
managed to send it before the crash. However, this should be OK
since Coordinator is aware of such duplicated messages.