fixup: Do not leave stale missingIssuer values behind
SSL_connect() may be called many times but it does not know about
missingIssuer and, hence, cannot clear that member before each call. We
have a choice of clearing missingIssuer before or after each call. Both
approaches appear to be equally (un)safe to me. The only difference is
that the latter is slightly cheaper because it can be conditional on the
flag being set (a condition that the caller has to check anyway).