Avoid using Security::SessionPointer when we can use the conn itself
Using pointers means that each function ought to handle nil values. The
associated repeated asserts and Musts are tedious but still unsafe.

Also add and use Security::Connection for TLS connections in new code.
The Connection alias allows us to avoid misleading, obscure, and long
SessionPointer::element_type (or raw SSL) and is, IMO, in line with the
2017 "switch session/connection for OpenSSL" discussions on squid-dev.