Eliminated IpcIoFile::HandleRebuildFinished()
and IORequestor::remoteIndexingCompleted() as well. This pair of methods
was activated in workers after receiving a message of an indexed disker.
Call a new Diskers::RemoteIndexingCompleted() instead.

I decided to change this after understanding that neither IpcIoFile no
its base DiskFile should receive these disker's 'indexed' notifications,
because, in general, a disker may include multiple 'files'.  Though
in rock there is a one-to-one dependency between disker and
IpcIoFile, this is an implementation detail we should not rely on.

Also avoid broadcasting premature 'indexed' notifications in Coordinator
from a not yet 'tagged' disker, which may be already registered but
still unavailable for workers.