fixup: Simplified PeerConnector::suspendedError_ maintenance
... and eased its upgrade path to ErrorDetails::Pointer.

PeerConnector does not (explicitly) uses TlsNegotiationDetails
refcounting, but refcounting simplifies the code a lot.
TlsNegotiationDetails will become ErrorDetails which is also refcounted.

Placing defaulted destructor in the .cc file is necessary to avoid
-Wdelete-incomplete errors that the default placement in the header file
would trigger. To generate the destructor, the compiler needs to know
full TlsNegotiationDetails declaration, and that knowledge is available
in .cc file only.