Assertion fix:  "assertion failed: ../../src/base/Lock.h:30: "count_ == 0""
This is inside Security::PeerConnector::resumeNegotiation while the
localy created TlsNegotiationDetails variable is destroyed at
method exit. This objst is set using the "=" operator which copies
between others the RefCountable::count member.
This patch uses a TlsNegotiationDetails::Pointer objects. I am expecting
this is to be a little faster in most cases too.