Implement the TODO: The caller interprets non-empty URIs as a sign of success
The Ssl::uriOfIssuerIfMissing method builds a list of URIs with missing
certificates but does not report if the URIs of one or more issuers
can not be retrieved. The callers of this function only checked if
the result provide urls from where someone can download 1-2 certificates,
but did not check if one or more issuer certificates can not be downloaded.
I suppose the reasons was the following:
  - The Ssl::uriOfIssuerIfMissing is faster in the case the issuer URI
    info is not reported in a certificate. This is not only in the case
    an issuer is missing but also in the case the issuer is a localy stored
    trusted CA certificate.
  - Probably the missing issuer certificate retrieved using  info
    attached on an other certificate or the certificate
    (which misses its issuer certificate) does not needed at all. This
    is breaks protocol rules, however a missing issuer certificate breaking
    the protocol too.
  - The drawback was that squid my download one or more certificates
    but in practice still can not verify the certificate because of
    missing issuers certificates but in most cases there is only one
    missing issuer certificate.

In any case the sites which have a missing issuer certificates are very few
and a not clear implementation does not worth the result.

This patch replaces the Ssl::uriOfIssuerIfMissing with a new function
Ssl::issuerIsMissing which runs a list of certificates and checks 1) if
the issuer certificate of the given certificate is not included in the
given certificates list and 2) if the issuer certificate does not
included in trusted or untrusted certificates list.
The callers uses this function to check if a certificate is missing
and use the existing Ssl::hasAuthorityInfoAccessCaIssuers to retrieve
URI info if exist.