Fix the case "successfull SSL_connect but missing issuer cert error"
Current branch code assumes that the SSL_connect, imediatelly after
the certificates are retrieved and verified, always should
result to a WANT_WRITE or WANT_READ error. But this is not true:
The SSL_connect internally, if wants to send data to remote server,
first try to write on socket and if fails to write all or some of the data
then result to a WANT_WRITE error.
However for the TLSv1.3, the SSL_connect after the server certificates
retrieved needs to write a small message (empty change-cipher-spec) and
it is very possible that the first try to write will succeed, and the
SSL_connect will return success.

This patch fixes branch code to handle the case we have a succeed
SSL_connect call, while we had detect an missing issuer certificate.
In this case downloads missing certificates, re-verify received
certificates and if all are OK, continue as normal.