Removed Ipc::StrandMessage::messageType as essentially unused
Side effect: Removed Ipc::TypedMsgHdr::type() as unused.

The kind of a StrandMessage is already hard-coded:

*  In message routing code: various "switch (rawType)" statements
*  In the StrandMessage sending code: SendMessage() calls in Coordinator
   and NotifyCoordinator() calls in other kids.

Repeating this hard-coded information in a class data member is useless.
This is why removing the messageType caused almost no code changes,
mostly moving the kind parameter from one method call to another.

The methods reacting to a routed StrandMessage are using the StrandCoord
information contained _inside_ the message. These methods do not double
check the message kind today. If we add those checks, we should migrate
to custom Message _types_ (e.g., based on a StrandMessageT<MessageKind>
template perhaps?) so that those added checks fail at _compile_ time.

TODO: I did not remove the single-field StrandMessage class because I
suspect we may add more fields to it soon. If not, we should remove it.