Do not rebuild doc/debug-* artifacts when we fail to scan all files
A growing number of srcFormat failures can be ignored with --keep-going.
Some failures might be (incorrectly) ignored even without that option by
old portions of srcFormat. We should not rebuild doc/debug-* files based
on a (potentially) _partial_ scan of sources. It is probably better to
keep those files as is than to corrupt them (in the presence of other
errors), complicating triage.