Replace defective Must2(c, "not c") API with Must3(c, "c", Here())
* We should not make a Must*() API different from the standard
  static_assert() API. It is just too confusing, especially since the
  two calls are very closely related.

* We should not tell Must*() writers to specify one condition (i.e. what
  should happen) but then describe the opposite condition (i.e. what
  went wrong). When the text describes the opposite condition, it is
  easy for a human writing or thinking about the text to type the wrong
  condition. For example, Must2(got < need, "incomplete message").

We should not keep the same macro name when changing the meaning of the
second parameter. Fortunately, adding a third argument to the macro fits
nicely into how modern Squid code should pass the source code location.

This commit represents the first attempt to fix the API. Like the
original Must2() macro, Must3() in this attempt supports SBuf condition
descriptions. IMO, this unfinished attempt started heading in the wrong
direction: It does not compile if quotes after XXX are removed, and it
requires a too-complex MakeFailedCheckMessage().