Check for too-large response headers earlier in processing. (#814)
If we are parsing the headers of a response, it is unnecessary to
process any of the headers if they are either too large, or their
length is unknown but the total response is greater than the configured
maxReplyHeaderSize. The debugging message for this case now features
a matching closing parenthesis.

This commit also removes an impossible-to-reach error in response
header processing, and documents that sanityCheckStartLine() returns
false for invalid headers.