fixup: Do not flush streams unless actually necessary
We do not want to imply that all "os << ..." code is missing a flush
command at the end. There are special cases where explicit flushing is
needed, but this is not one of them.

If flushing is needed here than PackableStream is buggy, but I hope that
is not the case because there were no XXXs added to (mark the bug and)
justify explicit flushing.