fixup: Fix fvdbDumpTable() profile, synced its name with branch changes
It does not make sense for fvdbDumpTable() to accept Packable just to
convert it to PackableStream -- we should not limit the number of
supported destination types unless we are gaining something from that
limit.

Currently, only two destination types makes sense:

* StoreEntry: Reduces caller code duplication.
* std::ostream: Increases dumping code usability (beyond StoreEntry).

I picked the former because I do not like code duplication and am not
sure we will eventually use non-entry dumping abilities.

I also added a paranoid check for safe StoreEntry dereference. I did not
move it into fvdbDumpCounts() to removing ourselves that we should fix
OBJH profile to use a StoreEntry reference so that the fvdbDumpVia()
_caller_ has to assure StoreEntry existence. The caller (direct or
indirect) will be able to handle missing StoreEntry much better than
this low-level code that should not worry about such things at all.
That OBJH fix is out of scope here due to its widespread impact.