fixup: Rename via_table to use modern CamelCase style
We are modifying all via_table lines anyway so this update does not
increase the diff. It will also make it a bit easier to find unused code
after forw_table is removed. There should be no "_table" lines left.

This change was missed when applying an earlier change request.