fixup: Sync proxySurrogate description with recent branch changes
The flag is not specific to HTTP ports. PROXY protocol is still not
supported on, say, ftp_port, but that limitation must be irrelevant to
flag users: Flag users must not assume a true value implies "HTTP".

There many other problems with TrafficMode flag descriptions, but we are
doing our best to stay focused on in-scope ones in hope to reduce review
overheads.