Fix ignore-cc and act-as-origin support for wildcard split-stack x_ports
The PortCfg::clone() hack (and clone_http_port_list() before it) forgot
to copy those two flags. This bug is fixed in this branch because it was
discovered while trying to remove the branch-added rawConfig() method
which was added specifically to work around clone() limitations.

Compilers will now be able to warn us if copying misses future members.

Also prohibited other forms of copying, nearly restricting copying to
the parsing code with special needs.