Removed TrafficMode::interceptedLocally()
We lack the resources necessary to explain to the reviewers why this
method improves the overall code quality.

No functionality changes compared to two commits ago (a7e9f64). Just
basic boolean logic manipulations to "inline" interceptedLocally().

Branch commit 6b979dd, which was trying to solve the same problem,
asserted that interceptedLocally() is true, but that assert is too
dangerous IMO -- the caller is not supposed to know that some of these
methods are unsafe to call in some contexts.