Bug 5029: assertion failed: MemBuf.cc: new_cap > (size_t) capacity
Since v4 the code converting URL-rewrite helper replies
from legacy Squid-2 syntax. Allocated a temporary MemBuf
sized to be exactly the response I/O size. This did not
allow for MemBuf::append use of grow() to ensure there
is always free space in the MemBuf and later code possibly
using terminate() to make the buffer a c-string.

Use the new MemBuf initialization to remove the max
capacity limitation on this temporary buffer and add
+1 to the initial size so the first allocation does not
need to grow() in order to terminate().