Cleanup HERE step 1 (#881)
Step 1:
 Add script to replace HERE symbol in a way that does
 not break branches during the remainder of the transition.

The script is added first as a separate commit for large
branches to merge and run before attempting to merge onto a
cleaned master branch.

The Enforcement commit is provided for review purposes to show
the changes made by the script. It should be removed before
merge.

Step 2:
 Master branch has a regular automated Enforcement commit.

Step 3:
 Then full removal of the HERE symbol as a third PR followup to
 drop it completely.