fixup: Do not imply that server_first code jumps over step2
That implication is currently correct -- the code does skip step2, but
only because step2 starts too late in the official code. We are supposed
to be inside step2 when we are parsing ClientHello... Moreover,
server_first is documented to not parse ClientHello.

Fixing this mess is out of scope here, but I think it is better not to
add to the mess with a comment implying that this code should not be
inside step2.

Also do not separate `http` computation from the function call it was
computed for. The step setting, on the other hand, is not just for
FwdState.