PeekingPeerConnector: Check the step once; do not quit if cannot check!
PeekingPeerConnector should only operate during step3, so we should
check the step when the job is created. If we ever add step4, it might
make sense to check that we are still at step3 while the job runs, but
that also depends on what step4 will be (and where it will be changed).

We could even check earlier, in FwdState constructor because no server
communication should happen in earlier steps. TODO?