Fix logging of client FQDNs in the default icap access.log format
The default ICAP access.log logformat has the equivalent of a %>A field.
Thus, the code must attempt to log FQDN regardless of the log_fqdn hack.

This fix has no effect in the current code because the log_fqdn value is
always true (due to other bugs), but fixing this code first makes fixing
those other bugs easier because it removes an anomaly in log_fqdn use.

TODO: The fix is not sufficient: Squid should also turn log_fqdn on when
using the default ICAP access.log format (a.k.a. "icap_squid").